Move class, enum to seperate file
