+++ src/models/room/ChatMessage.ts
@@ -1,0 +1,5 @@
+export interface ChatMessage {
+  readonly id: string;
+  readonly authorName: string;
+  readonly content: string;
+}
