+++ src/models/room/RoomState.ts
@@ -1,0 +1,5 @@
+export enum RoomState {
+  CREATED,
+  CONNECTED,
+  JOINED,
+}
